+++ MWSE/TES3Cell.h
@@ -1,5 +1,6 @@
 #pragma once
 
+#include "TES3CellFlags.h"
 #include "TES3Defines.h"
 
 #include "TES3Object.h"
@@ -8,20 +9,6 @@
 #include "NISourceTexture.h"
 
 namespace TES3 {
-	namespace CellFlag {
-		typedef unsigned int value_type;
-
-		enum Flag : value_type {
-			Interior = 0x1,
-			HasWater = 0x2,
-			SleepIsIllegal = 0x4,
-			WasLoaded = 0x8,
-			TempRefsLoaded = 0x10,
-			MarkerDrawn = 0x20,
-			BehavesAsExterior = 0x80
-		};
-	}
-
 	struct MapNote {
 		NI::Point3 position; // 0x0
 		char * text; // 0xC
